finalized new C++ backend
