+++ src/polyhedralGravityPython/python-interface.cpp
@@ -12,30 +12,6 @@
 
 namespace py = pybind11;
 
-/**
- * This method converts a named tuples GravityModelResult to a tuple
- * @param result a named tuple of type GravityModelResult
- * @return tuple of potential, acceleration and gradiometric tensor
- *
- * TODO: Remove this
- */
-std::tuple<double, std::array<double, 3>, std::array<double, 6>> convertToTuple(
-        const polyhedralGravity::GravityModelResult &result) {
-    return result;
-}
-
-/**
- * This method converts a vector of named tuples GravityModelResult to a vector of tuples
- * @param resultVector a vector of named tuples of type GravityModelResult
- * @return vector of tuples of potential, acceleration and gradiometric tensor
- *
- * TODO: Remove this
- */
-std::vector<std::tuple<double, std::array<double, 3>, std::array<double, 6>>> convertToTupleVector(
-        const std::vector<polyhedralGravity::GravityModelResult> &resultVector) {
-    return resultVector;
-}
-
 PYBIND11_MODULE(polyhedral_gravity, m) {
     using namespace polyhedralGravity;
     m.doc() = "Computes the full gravity tensor for a given constant density polyhedron which consists of some "
